requirements.txt
streamlit
vertexai
google-cloud-aiplatform
python-dotenv